받은 id 배열에 해당되는 게시물에 종속된 사진 데이터에 게시물 데이터 추가

+++ src/main/java/com/example/fileupload/article/dto/CreateArticleForm.java
@@ -19,5 +19,5 @@ public class CreateArticleForm {
     @Size(min = 1,max = 3000)
     private String body;
 
-    private List<Integer> imageIdList;
+    private List<Long> imageIdList;
 }
